backend/requirements.txt
Flask==2.2.3
Werkzeug==2.3.8
gunicorn==20.1.0
numpy==1.23.5
scikit-learn==1.2.2
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.11.0
joblib
protobuf==3.20.3